+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Allow orga to apply for working groups !1953 #1050 @alex.simm
 - Excluded wall post author from store team members notified about new post !1960 @fabian.rudolf
 - Business Card generation: show info and cut of the rest if street or plz + city is longer than 49 characters #834 !1489 @treee111
+- Fix permissions for content IDs @alex.simm
 
 ## Refactoring
 - Change DTOs for dashboard updates to use date objects !1926 @alex.simm